Rename the "descript" module to "describe"
