+++ printo/__init__.py
@@ -1,7 +1,7 @@
-from printo.descript import (
+from printo.describe import (
     describe_data_object as describe_data_object,
 )
-from printo.descript import (
+from printo.describe import (
     descript_data_object as descript_data_object,
 )
 from printo.filters import not_none as not_none
