+++ printo/describe.py
